Make return types on gnlocs, nlocs, nvars and nrecs consistent in ObsSpace Fortran interface (#423)
* Made gnlocs, nlocs, nvars, nrecs "integer(kind=c_size_t)" type in the Fortran interface so that these
are consistent with the C++ types.

* Decided to do conversion to integer (from c_size_t) at Fortran module level since there
are many places in UFO to switch from integer to c_size_t and the conversion from c_size_t to
integer is safe for these cases.

+++ src/ioda/obsspace_interface.f
@@ -29,28 +29,28 @@ type(c_ptr) function c_obsspace_obsvariables(obss) bind(C, name='obsspace_obsvar
   type(c_ptr), value :: obss
 end function c_obsspace_obsvariables
 
-integer(kind=c_int) function c_obsspace_get_gnlocs(obss) bind(C,name='obsspace_get_gnlocs_f')
+integer(kind=c_size_t) function c_obsspace_get_gnlocs(obss) bind(C,name='obsspace_get_gnlocs_f')
   use, intrinsic :: iso_c_binding
   implicit none
 
   type(c_ptr), value :: obss
 end function c_obsspace_get_gnlocs
 
-integer(kind=c_int) function c_obsspace_get_nlocs(obss) bind(C,name='obsspace_get_nlocs_f')
+integer(kind=c_size_t) function c_obsspace_get_nlocs(obss) bind(C,name='obsspace_get_nlocs_f')
   use, intrinsic :: iso_c_binding
   implicit none
 
   type(c_ptr), value :: obss
 end function c_obsspace_get_nlocs
 
-integer(kind=c_int) function c_obsspace_get_nrecs(obss) bind(C,name='obsspace_get_nrecs_f')
+integer(kind=c_size_t) function c_obsspace_get_nrecs(obss) bind(C,name='obsspace_get_nrecs_f')
   use, intrinsic :: iso_c_binding
   implicit none
 
   type(c_ptr), value :: obss
 end function c_obsspace_get_nrecs
 
-integer(kind=c_int) function c_obsspace_get_nvars(obss) bind(C,name='obsspace_get_nvars_f')
+integer(kind=c_size_t) function c_obsspace_get_nvars(obss) bind(C,name='obsspace_get_nvars_f')
   use, intrinsic :: iso_c_binding
   implicit none
